wip on selectclasses
