+++ study_group_app/lib/screens/profile/profile_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:study_group_app/models/user.dart';
 import 'package:study_group_app/screens/profile/settings_page.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:study_group_app/screens/student/add_courses.dart';
 import 'package:study_group_app/screens/student/courses.dart';
 import 'package:study_group_app/services/auth.dart';
 
@@ -18,55 +19,62 @@ class Profile extends StatelessWidget {
         title: Text('Profile'),
         backgroundColor: Theme.of(context).appBarTheme.color,
         elevation: 0.0,
-        actions: <Widget> [
+        actions: <Widget>[
           IconButton(
-            icon: Icon(Icons.more_vert, size: 40.0), 
+            icon: Icon(Icons.more_vert, size: 40.0),
             onPressed: () {
               Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => ProfileSettings(uid: user.uid, auth: _auth,)));
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => ProfileSettings(
+                            uid: user.uid,
+                            auth: _auth,
+                          )));
             },
           )
         ],
       ),
-      backgroundColor: Theme.of(context).backgroundColor, 
+      backgroundColor: Theme.of(context).backgroundColor,
 
       body: ListView(
         children: <Widget>[
-          SizedBox(height: 50.0,),
-
+          SizedBox(
+            height: 50.0,
+          ),
           Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-
               Container(
                 height: 125.0,
                 width: 125.0,
                 child: CircleAvatar(
                   radius: 60,
                   backgroundColor: Colors.grey.shade300,
-                  child: FaIcon(
-                    FontAwesomeIcons.solidUser,
-                    size: 90, 
-                    color: Colors.white
-                  ),
+                  child: FaIcon(FontAwesomeIcons.solidUser,
+                      size: 90, color: Colors.white),
                 ),
               ),
 
-              SizedBox(height: 25.0,),
+              SizedBox(
+                height: 25.0,
+              ),
 
               // Retrieve username from Firestore
-              Text('${user.firstName} ${user.lastName}',
+              Text(
+                '${user.firstName} ${user.lastName}',
                 style: TextStyle(
                   color: Colors.black,
                   fontSize: 24,
                 ),
               ),
 
-              SizedBox(height: 20.0,),
+              SizedBox(
+                height: 20.0,
+              ),
 
               // Default school
-              Text('California State University, Chico',
+              Text(
+                'California State University, Chico',
                 style: TextStyle(
                   color: Colors.black,
                   fontSize: 20,
@@ -80,13 +88,18 @@ class Profile extends StatelessWidget {
                   onPressed: () {
                     Navigator.push(
                       context,
-                      MaterialPageRoute(builder: (context) => CourseViewerState()));
+                      MaterialPageRoute(
+                        builder: (context) => user.courses.isEmpty
+                            ? AddCourses()
+                            : CourseViewerState(),
+                      ),
+                    );
                   },
                   shape: RoundedRectangleBorder(
                     borderRadius: BorderRadius.circular(30),
                   ),
                   child: Text(
-                    'CLASSES',
+                    user.courses.isEmpty ? 'ADD CLASSES' : 'VIEW CLASSES',
                     style: TextStyle(
                       color: Color(0xFF98c1d9),
                       letterSpacing: 1.5,
